+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 * Added GALI for continuous systems.
 * Added the Henon Helies system in famous_systems. TODO is still to add
   callbacks that conserve energy.
+* Bugfix of the printing of `Dataset` on Juno.
+* Added `BigDiscreteDS` (still need methods for e.g. lyapunovs).
 
 # v0.6.0:
 * Many minor bugfixes and many corrections to the documentation examples.
@@ -9,7 +11,8 @@
   interaction with NearestNeighbors.jl
 * **[BREAKING]** : `timeseries` function was renamed to `trajectory`.
     * Now the documentation is clear: `trajectory` means a set of one-dimensional
-      timeseries (or a *multi-dimensional* timeseries), while `timeseries` always means a one-dimensional timeseries
+      timeseries (or a *multi-dimensional* timeseries), while `timeseries`
+      always means a one-dimensional timeseries
 * Removed function `perform_non0hist`, as it had no reason to exist.